Update 2.0.2
I've added the first two logical operators "AND" and "OR" (The "AND" operator does not make that much because a white space is always an "AND", but.. :D)

You can use the "OR" operator to display resources with different keywords at the same time.

eg.
"TimeVote OR WeatherVote"
"HungerGames OR SurvivalGames"

I will add some more in the future. :)